perf(amd64): pin integer locals in registers (Valent-Block) (#25)
* perf(amd64): pin integer locals in registers (Valent-Block)

Port WARP's locals-in-registers optimization: the first few integer
locals (params first, the hottest) are pinned to dedicated registers
(RBX/R13/R14/R15) instead of living in frame slots, so local.get/set/tee
avoid memory traffic and operands fold the register directly.

Design (faithful in spirit; v1 scoped for low risk):
- New lazy operand kind vPinned: local.get of a pinned local pushes a
  reference to its register (no code); consumers fold it as an RHS, and
  it is copied (never reused as a mutable dest) when materialized.
- The pinned registers are removed from the scratch pool (cg.reserved),
  so they are never handed out as temporaries.
- Canonical-slot joins need no extra work: a pinned local always lives
  in the same register on every edge into a join, so branches/loops stay
  consistent for free; flush() copies the register to the operand slot
  but leaves the local in place.
- Calls clobber the registers (they are plain scratch in the callee), so
  emitWrapperCall spills the pinned locals to their frame slots around
  the call and reloads after. Host calls (deferred log) don't clobber.
- materializeLocalRefs captures pending vPinned reads before local.set
  overwrites the register (the lazy-aliasing hazard).
- Float locals stay frame-resident (integer-only in v1).

Tests: new pinned_test.go covers the >pool-size locals boundary (pinned
+ frame-resident mixed in one expr), double-use, alias capture, loop
back-edge consistency, and if/else join; existing recursion/AS payloads
exercise pinned params across calls. Full `go test ./...` green incl.
-race; CLI spot-checks fib(25)=75025, sumto(100)=5050.

Perf (bench/ Exec suite, count=8, benchstat) — geomean -17.1% sec/op:

  Exec/arith.run          4.26us -> 1.48us   -65.3%
  Exec/fib_iter.fib       65.8ns -> 35.6ns   -45.8%
  Exec/memory.sum          821ns ->  462ns   -43.8%
  ExecFibLoop_wago        34.7ns -> 21.4ns   -38.4%  (wazero 43.7ns)
  Exec/globals.accumulate 2.48us -> 1.76us   -29.1%
  Exec/tiny.add           26.7ns -> 25.1ns    -6.1%

Regressions are confined to call-heavy paths, the known cost of spilling
pinned locals around calls (option A): ExecFibRec_wago +3.6%,
Exec/dispatch.apply (call_indirect) +7.5%. A register-based internal
call ABI (a later port) is the natural fix; exec allocs unchanged (0).

* test(amd64): add module runner helper for internal call tests

* test(amd64): cover pinned locals across internal calls

* test(amd64): cover multiple pending pinned local reads

* test(amd64): cover pinned locals through br_table

* docs(amd64): clarify pinned register ABI assumptions

+++ src/core/compiler/backend/amd64/localfuse_test.go
@@ -83,11 +83,14 @@ func TestLocalSetGetCorrectness(t *testing.T) {
 // register. We check there is no `mov reg,[rbp-...]` that targets the same slot
 // a preceding store wrote — simplest proxy: the fused function is strictly
 // smaller than the same logic with a non-fusing barrier in between.
+//
+// Uses f32 locals, which are frame-resident: integer locals are pinned in
+// registers, where local.get is already free and the peephole is a no-op.
 func TestLocalSetGetShrinksCode(t *testing.T) {
-	fused := watToModule(t, `(module (func (export "f") (param i32) (result i32) (local i32)
-		local.get 0 local.get 1 i32.add local.set 1 local.get 1))`)
-	notFused := watToModule(t, `(module (func (export "f") (param i32) (result i32) (local i32)
-		local.get 0 local.get 1 i32.add local.set 1 i32.const 0 drop local.get 1))`)
+	fused := watToModule(t, `(module (func (export "f") (param f32) (result f32) (local f32)
+		local.get 0 local.get 1 f32.add local.set 1 local.get 1))`)
+	notFused := watToModule(t, `(module (func (export "f") (param f32) (result f32) (local f32)
+		local.get 0 local.get 1 f32.add local.set 1 i32.const 0 drop local.get 1))`)
 	fc, err := CompileFunction(fused, 0)
 	if err != nil {
 		t.Fatal(err)
